requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Login
python-dotenv
Werkzeug
pandas
openpyxl
reportlab
